requirements.txt
amqp==2.4.2
apipkg==1.5
appdirs==1.4.3
appnope==0.1.0
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==18.2.0
autobahn==19.3.3
Automat==0.7.0
autopep8==1.4.2
backcall==0.1.0
beautifulsoup4==4.6.0
billiard==3.6.0.0
black==18.9b0
celery==4.3.0
certifi==2018.10.15
cffi==1.12.2
channels==2.1.7
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
daphne==2.2.5
decorator==4.3.2
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2
django-admin==1.3.2
django-allauth==0.38.0
django-appconf==1.0.3
django-cogwheels==0.2
django-crispy-forms==1.7.2
django-excel-response2==2.0.8
django-fernet-fields==0.5
django-filter==2.1.0
django-guardian==1.5.0
django-jsonfield==1.1.0
django-modelcluster==4.3
django-redis==4.10.0
django-rest-swagger==2.2.0
django-six==1.0.4
django-taggit==0.23.0
django-treebeard==4.3
djangorestframework==3.9.2
djangorestframework-xml==1.4.0
djangorestframework-yaml==1.0.3
draftjs-exporter==2.1.5
execnet==1.6.0
flake8==3.6.0
future==0.17.1
greenlet==0.4.15
gunicorn==19.9.0
h11==0.8.1
html5lib==1.0.1
httptools==0.0.13
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipython==7.3.0
ipython-genutils==0.2.0
isort==4.3.4
itypes==1.1.0
jedi==0.13.2
Jinja2==2.10.1
kombu==4.5.0
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==7.0.0
oauthlib==2.1.0
openapi-codec==1.3.2
OWSLib==0.17.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.9.0
portalocker==1.4.0
prompt-toolkit==2.0.8
psycopg2-binary==2.7.7
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
Pygments==2.3.1
PyHamcrest==1.9.0
pyperclip==1.7.0
pyproj==1.9.6
pytest==4.4.1
pytest-django==3.4.8
pytest-forked==1.0.2
pytest-xdist==1.28.0
python-dateutil==2.8.0
python-dotenv==0.10.1
python3-openid==3.1.0
pytz==2018.7
PyYAML==5.1
redis==3.2.1
requests==2.20.1
requests-oauthlib==1.0.0
rest-framework-generic-relations==1.2.1
screen==1.0.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
traitlets==4.3.2
travis-encrypt==1.2.3
Twisted==18.9.0
txaio==18.8.1
tzlocal==1.5.1
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.2
uvicorn==0.6.1
vine==1.3.0
wagtail==2.4
wagtail-blocks==0.6.8
wagtailfontawesome==1.1.4
wagtailmenus==2.12
wcwidth==0.1.7
webencodings==0.5.1
websockets==7.0
Willow==1.1
xlwt==1.3.0
zope.interface==4.6.0
